// apps/api/src/resources/api-messages.ts
export enum ApiMessages {
  RECORD_NOT_FOUND = 'Запись в сводке не найдена',
  BASE_ALREADY_ON_CHECK = 'Основа уже отнесена на пробу',
  PRODUCT_ALREADY_ON_CHECK = 'Продукт уже отнесен на пробу',
  NEED_EMPTY_OR_FAIL_OR_CORRECT = 'Необходимо отсутствие записей или статусы "Брак продукта" или "Требуется доработка"',
  NEED_PROGRESS_OR_FAIL_OR_CORRECT = 'Необходимы статусы "Фасуется", "Брак продукта" или "Требуется доработка"',
  NEED_PASS = 'Необходимо статус "Допуск на подключение"',
  SECOND_CHECK_FAIL = 'Для повторной пробы необходим статус "Брак основы", "Требуется корректировка" или "Продолжение варки"',
}